==> 4_figures_immune.ipynb <==
# Figures associated to immune processing

**You need to tun the notebook 1_ first!!!!**

from cellassign import assign_cats
import gzip
import itertools as itl
import matplotlib.pyplot as plt
import matplotlib as mpl
import numpy as np
import os
import pandas as pd
import scanpy as sc
import scanpy.external as sce
import seaborn as sns
import triku as tk
import itertools
import matplotlib.pylab as pylab
import matplotlib.patches as mpatches

from functions import adata_plot_KOvsWT, stat_annot_gene, save_adata
from functions import dict_WT_KO_colors

from datetime import date

today = date.today()
today = today.strftime("%Y_%m_%d")

# Palettes for UMAP gene expression

magma = [plt.get_cmap('magma')(i) for i in np.linspace(0,1, 80)]
magma[0] = (0.88, 0.88, 0.88, 1)
magma = mpl.colors.LinearSegmentedColormap.from_list("", magma[:65])

params = {'figure.dpi': 200,
          'legend.fontsize': 'large',
         'axes.labelsize': 'x-large',
         'axes.titlesize':'x-large',
         'xtick.labelsize':'x-large',
         'ytick.labelsize':'x-large', 
         'axes.grid': False}
pylab.rcParams.update(params)
# mpl.style.use('default')

pd.set_option('display.max_columns', None)

os.makedirs('results/figures/F3', exist_ok=True)

## Functions and variables

%store -r dict_WT_KO_colors
%store -r seed

## Load adata

adata_all_harmony = sc.read_h5ad('adatas/adata_all_harmony.h5')
adata_all_harmony.uns['log1p'] = {'base': None}  # Becuase sometimes it fails at DEG calculation
adata_imm = adata_all_harmony[adata_all_harmony.obs['merged_cell_type_name'].isin(['APC', 'LYM', 'NEU', 'MC'])]

sc.pp.filter_genes(adata_imm, min_counts=1)

sc.pp.pca(adata_imm, random_state=seed, n_comps=50)
sce.pp.harmony_integrate(adata_imm, key='batch', max_iter_harmony=50, plot_convergence=True)
sc.pp.neighbors(adata_imm, random_state=seed, n_neighbors=int(len(adata_imm) ** 0.5 * 0.5), metric='cosine', use_rep='X_pca_harmony')
tk.tl.triku(adata_imm, use_raw=False)

sc.pp.pca(adata_imm, random_state=seed, n_comps=50)
sce.pp.harmony_integrate(adata_imm, key='batch', max_iter_harmony=50, plot_convergence=True)
sc.pp.neighbors(adata_imm, random_state=seed, n_neighbors=int(len(adata_imm) ** 0.5 * 0.5), metric='cosine', use_rep='X_pca_harmony')

sc.tl.umap(adata_imm, min_dist=1)

### Subpopulation annotation

To do the annotation we used the characterization made by Joost et al. 2020. There are some populations that are not appearing (Krt75+ to Krt18+) and we are not sure about their functional relationship

dict_immune = {  '0: αβT': ['Tnfrsf18', 'Cxcr6', 'Ccr2', 'Ifngr1', 'Cd52', 'Rora', 'Il7r', 'Aebp2', 'Icos', 'Dgat1', 'Shisa5', 'Grap', 'Cd27', 'Pdcd1', 'Ikzf3', 'Il18r1', 
                             'Gata3', 'Spn', 'Tnfrsf4'],
                 '1: γδT': ['Nkg7', 'Ctsw', 'Cd7', 'Rab4a', 'Ctla2a', 'Rgs2', 'Cd3d', 'Cd3g', 'Fermt2', 'Tcrg-C1', 'Klrk1', 'Trf', 'Itk', 'Ubald2', 'Il2rb', 'Map1b', 'Klrb1b', 'Cd2', 
                              'Trbc2', 'Abcb1b', 'Rasl11b', 'Parp8', 'Tiam1'],
                 '2: Act. Neu': ['Acod1', 'Il1f9', 'Csta3', 'Rhov', 'Stfa2l1', 'Ccl3', 'Hcar2', 'Slpi', 'Ccrl2', 'Il1rn', 'Cxcl3', 'G0s2', 
                                      'Stfa2', 'Ptgs2', 'Asprv1', 'Fpr1'],
                 '3: Neu': ['Camp', 'Ltf', 'Chil3', 'S100a8', 'S100a9', 'Tinagl1', 'Stfa1', 'Ngp', 'Chil3', 'Cd177', 'Adpgk', 'Syne1', 'Cebpe', 'Hsd11b1', 'Mgst1', 
                                 'Tst', 'Ly6c2', 'Pilrb2', 'Fcnb', 'Rab44', 'Nfe2', 'Cpne3', 'Ldhc', 'Hmgn2', 'Adgrg3'],
                 '4: Act. LC': ['Fscn1', 'Cd200', 'Il4i1', 'Pkib', 'Wnk1', 'Cd74', 'Cst3', 'Ccr7', 'Bcl2a1a', 'Cacnb3', 'Ccl22', 'Gm13546', 'Serpinb9b', 'Relb', 'Cd40', 'Arl5a'],
                 '5: LC': ['Mfge8', 'Cd207', 'Cldn1', 'Pgf', 'Cd74', 'H2-Aa', 'H2-Eb1',  'Nr4a3', 'Il1r2', 'Pkib', 'Grk3', 'Ktn1', 'Pxdc1', 'Dnah2', 'Timd4', 
                                      'Epcam', 'Ccr7', 'St14', 'Mreg', 'Cd40'],
                 '6: Mφ/DCA': ['Nos2', 'Saa3', 'Arg1', 'Wfdc17', 'Csf1r', 'F10', 'Ly6c2', 'Gsr', 'Anxa5', 'Ptgs2', 'Cxcl3', 'Pid1', 'Serpinb2', 'Csf3', 'Vcan', 'Inhba', 'Ptges', 
                               'F13a1', 'Met', ],
                 '7: Mφ/DCB': ['Pf4', 'C1qb', 'Igf1', 'Grn', 'Dab2', 'C1qc', 'Apoe', 'C1qa', 'Ms4a7', 'Timp2', 'Cd36', 'Mrc1', 'Fxyd2', 'F13a1', 'Rgl1', 'Ednrb', 'Hgsnat', 'Tgfbi', 'Rnase4', 
                                 'Sash1', 'Folr2', 'Gas6', 'Stab1', 'Ccl8', 'Fcgrt', 'Ms4a6c', 'P2rx7'] ,
                 '8: Mφ/DCC': ['Cd209a', 'Cd209d', 'Slc39a14', 'Plbd1', 'Aif1', 'Cd74', 'Clec4e', 'Tgfbi', 'Pdpn', 'Cxcl3', 'Vat1', 'Eps8', 'Tppp3', 
                                     'Procr', 'Tarm1', 'Clec4a3', 'Ccl5'],
                 '9: MC': ['Cma1', 'Tpsb2', 'Mcpt4', 'Fdx1', 'Hs3st1', 'Cd55', 'Tpsab1', 'Hdc', 'Tph1', 'Cpa3', 'Fgfr1', 'Gata2', 'Slco2b1', 'Vwa5a', 'Il1rl1', 
                                'Cyp11a1', 'Adora3', 'Kit', 'Ap1s3', 'Hexb', 'Creb3l1', 'Mrgprb1', 'Rab27b', 'Penk', 'Hs6st2', 'Adamts9', 'Akap12', 'Slc45a3', 'Mrgprb2', 'Gnai1'], 
                 '10: Bas': ['Mcpt8', 'Cyp11a1', 'Cd200r3', 'Ifitm1', 'Ero1l', 'Ccl6', 'Ccl3', 'Lilrb4a', 'Gata2', 'Ccl4', 'Hdc', 'Osm', 'Lat2', 'Tec', 'Rab44', 'Csf2rb2', 
                               'Os9', 'Ier3', 'Aqp9', 'Cd200r4', 'Emilin2', 'Csf1', 'Slpi', 'Hmox1', 'Mboat1', 'Hs3st1', 'Adgrg3', ],
                 
                 '11: BC': ['Vpreb3', 'Cd79a', 'Cd79b', 'Chchd10', 'Ighm',  'Stmn1', 'Igll1', 'Dut', 'Top2a', 'Rasgrp2', 'Myb', 'Ssrp1', 'Gimap6', 'Selenoh', 'Apobec3', 'Smc2',
                             'Cks1b', 'Lgals9', 'Ebf1', 'Csrp2', 'Adgrg1', 'Pou2af1', 'Tcf3', 'Hist1h1a', 'Parp1', 'Lig1', 'Dntt', 'Sox4'],
               
                 '12: PC': ['Iglc3', 'Ms4a1', 'Iglc2', 'Ighm', 'Igkc', 'Cd79a', 'Ly6d', 'Cd79b', 'Tnfrsf13c', 'Iglc1', 'Sdc1', 'Cd74', 
                                  'Txndc16', 'Ighd'],    
                 
                                }


dict_imm_colors = { '0: αβT': '#0086b9',
                    '1: γδT': '#99DCF5',
                    '2: Act. Neu': '#D3BC5F',
                    '3: Neu': '#A5901A',
                    '4: Act. LC': '#8E0C6F',
                    '5: LC': '#D40055',
                    '6: Mφ/DCA': '#AE54B8',
                    '7: Mφ/DCB': '#4D0765',
                    '8: Mφ/DCC': '#FF5599',
                    '9: MC': '#FF9955',
                    '10: Bas': '#A05A2C',
                    '11: BC': '#4c708a',
                    '12: PC': '#00171F',}


sc.tl.leiden(adata_imm, resolution=4)
assign_cats(adata_imm, dict_cats=dict_immune, min_score=0.4, quantile_gene_sel=0.6, key_added='subtype')

adata_imm.obs['subtype_name'] = [i.split(': ')[1] for i in adata_imm.obs['subtype']]
adata_imm.obs['subtype_name'] = adata_imm.obs['subtype_name'].astype('category')
adata_imm.obs['subtype_name'].cat.reorder_categories([i.split(': ')[1] for i in dict_imm_colors.keys()], inplace=True)

adata_imm.obs['subtype_number'] = [i.split(': ')[0] for i in adata_imm.obs['subtype']]
adata_imm.obs['subtype_number'] = adata_imm.obs['subtype_number'].astype('category')
adata_imm.obs['subtype_number'].cat.reorder_categories([i.split(': ')[0] for i in dict_imm_colors.keys()], inplace=True)


adata_imm.uns['subtype_colors'] = list(dict_imm_colors.values())
adata_imm.uns['subtype_name_colors'] = list(dict_imm_colors.values())
adata_imm.uns['subtype_number_colors'] = list(dict_imm_colors.values())

## Merge some subtypes to make general types

dict_changes = {'2: Act. Neu': '2/3: Neu',
                '3: Neu': '2/3: Neu',
                '4: Act. LC': '4/5: LC',
                '5: LC': '4/5: LC',
                '6: Mφ/DCA': '6/7/8: Mφ/DC',
                '7: Mφ/DCB': '6/7/8: Mφ/DC',
                '8: Mφ/DCC': '6/7/8: Mφ/DC',
                }


dict_imm__merge_colors = {  '0: αβT': '#0086b9',
                            '1: γδT': '#99DCF5',
                            '2/3: Neu': '#D3BC5F',
                            '4/5: LC': '#D40055',
                            '6/7/8: Mφ/DC': '#AE54B8',
                            '9: MC': '#FF9955',
                            '10: Bas': '#A05A2C',
                            '11: BC': '#4c708a',
                            '12: PC': '#00171F',}

adata_imm.obs['subtype_merged'] = [dict_changes[i] if i in dict_changes.keys() else i for i in adata_imm.obs['subtype']]


adata_imm.obs['subtype_merged_name'] = [i.split(': ')[1] for i in adata_imm.obs['subtype_merged']]
adata_imm.obs['subtype_merged_name'] = adata_imm.obs['subtype_merged_name'].astype('category')
adata_imm.obs['subtype_merged_name'].cat.reorder_categories([i.split(': ')[1] for i in dict_imm__merge_colors.keys()], inplace=True)

adata_imm.obs['subtype_merged_number'] = [i.split(': ')[0] for i in adata_imm.obs['subtype_merged']]
adata_imm.obs['subtype_merged_number'] = adata_imm.obs['subtype_merged_number'].astype('category')
adata_imm.obs['subtype_merged_number'].cat.reorder_categories([i.split(': ')[0] for i in dict_imm__merge_colors.keys()], inplace=True)


adata_imm.uns['subtype_merged_colors'] = list(dict_imm__merge_colors.values())
adata_imm.uns['subtype_merged_name_colors'] = list(dict_imm__merge_colors.values())
adata_imm.uns['subtype_merged_number_colors'] = list(dict_imm__merge_colors.values())

## Table with markers and p-values

sc.tl.rank_genes_groups(adata_imm, groupby='subtype')
names = pd.DataFrame(adata_imm.uns['rank_genes_groups']['names'])
pvals = pd.DataFrame(adata_imm.uns['rank_genes_groups']['pvals_adj'])

for cell_type, list_cell_markers in dict_immune.items():
    print(cell_type)
    sc.pl.umap(adata_imm, color=[i for i in list_cell_markers if i in adata_imm.var_names], alpha=0.5, ncols=5, cmap=magma, use_raw=False)

# Create a DataFrame with 30 most important genes
list_dfs = []
for subtype, genes in dict_immune.items():
    df = pd.DataFrame({f'{subtype} genes': names[subtype], f'p-value {subtype}': pvals[subtype]}, dtype=object)
    df_short = df[df[f'{subtype} genes'].isin(genes)].sort_values(by=f'p-value {subtype}', ascending=True).iloc[:30].reset_index(drop=True)
    df[f'p-value {subtype}'] = df[f'p-value {subtype}'].apply(lambda x: '%.3E' % x).values.tolist()
    list_dfs.append(df)

df_all_celltypes = pd.concat(list_dfs, axis=1).iloc[:50]
df_all_celltypes.to_csv(f'results/{date.today().strftime("%Y_%m_%d")}-IMM_DEGs.csv.csv', index=False)
df_all_celltypes

## UMAP of fibroblast subpopulations and batch

#### Supp Fig to Fig 3

adata_imm_WT = adata_imm[adata_imm.obs['batch'].isin(['WT1', 'WT2'])]
adata_imm_WT.uns['batch_colors'] = [dict_WT_KO_colors['WT1'], dict_WT_KO_colors['WT2'], ]
adata_imm_KO = adata_imm[adata_imm.obs['batch'].isin(['KO1', 'KO2'])]
adata_imm_KO.uns['batch_colors'] = [dict_WT_KO_colors['KO1'], dict_WT_KO_colors['KO2'], ]

fig, axs = plt.subplots(1, 1, figsize=(8, 4))
s = 18

sc.pl.umap(adata_imm, color='subtype_number', s=s, ax = axs, 
           legend_loc='on data', frameon=False, show=False, 
           legend_fontoutline=4, legend_fontsize=14)
sc.pl.umap(adata_imm, color='subtype', s=s, ax = axs, 
           frameon=False, show=False)

title_obj = axs.set_title('') 

plt.tight_layout()

plt.savefig(f'results/figures/F3/makers_imm_closeup_{date.today().strftime("%Y_%m_%d")}.png', dpi=300)
plt.savefig(f'results/figures/F3/Supp_Fig3A_closeup_{date.today().strftime("%Y_%m_%d")}.png', dpi=300)

fig, axs = plt.subplots(2, 2, figsize=(12, 7))
s = 18


sc.pl.umap(adata_imm_WT, color='batch', s=s, ax = axs.ravel()[0], 
           frameon=False, show=False, na_in_legend=False)


sc.pl.umap(adata_imm_WT, color='subtype_number', s=s, ax=axs.ravel()[1], 
           legend_loc='on data', frameon=False, show=False, legend_fontoutline=4, legend_fontsize=14, 
           na_in_legend=False, title='condition')
sc.pl.umap(adata_imm_WT, color='subtype', s=s, ax = axs.ravel()[1], 
           frameon=False, show=False, na_in_legend=False, title='FB type')



sc.pl.umap(adata_imm_KO, color='batch', s=s, ax = axs.ravel()[2], 
           frameon=False, show=False, na_in_legend=False, title='')


sc.pl.umap(adata_imm_KO, color='subtype_number', s=s, ax=axs.ravel()[3], 
           legend_loc='on data', frameon=False, show=False, legend_fontoutline=4, legend_fontsize=14, 
           na_in_legend=False, title='')


plt.tight_layout()

plt.savefig(f'results/figures/F3/makers_imm_{date.today().strftime("%Y_%m_%d")}.png', dpi=300)
plt.savefig(f'results/figures/F3/Supp_Fig3A_{date.today().strftime("%Y_%m_%d")}.png', dpi=300)

## log2fold heatmap highlighting the expression of specific markers

markers = {'Suppressive': ['Cd274', 'Pdcd1lg2', 'Cd80', 'Lgals9', 'Lgals3', 'Mif', 'Cd74', 
                          'Adora2a', 'Adora2b', 'Arg1', 'Arg2', 'Il4i1',],  
           'Stimulatory': ['Il10', 'Tnfsf9', 'Tnfsf4', 'Cd86',], 
           'Proinflammatory cytokines': ['Il1a', 'Il1b', 'Il1rn', 'Tnf', 'Il6', 'Ifng', 
              'Tgfb1',],
           'IL1 processing': ['Nlrp3', 'Casp1', 'Casp8', 'Gzmb', ], 
           'Chemoattractant': ['Ccl3', 'Ccl4', 'Ccl5', 'Ccl17', 'Ccl22', 'Cxcl1', 
                                'Cxcl2', 'Cxcl3', 'Cxcl16', 'Csf3', 'Csf3r'],
           'Chemoreceptors': ['Ccr1', 'Ccr2', 'Ccr4', 'Ccr7', 'Cxcr6', 'C3ar1', 'Il7r', 'Il2ra']
           
          }

list_imm_logfold_genes = [j for i in markers.values() for j in i]

categories_order = adata_imm.obs['subtype_merged_name'].cat.categories
categories_order = [i for i in categories_order if i != 'Bas']
# There are no basophil cells in WT, so the logfold is 0

df_logfold = pd.DataFrame(columns=categories_order, index=list_imm_logfold_genes)
df_mean = pd.DataFrame(columns=categories_order, index=list_imm_logfold_genes)

for cell_type in categories_order:
    adata_WT = adata_imm_WT[adata_imm_WT.obs['subtype_merged_name'] == f"{cell_type}"]
    adata_KO = adata_imm_KO[adata_imm_KO.obs['subtype_merged_name'] == f"{cell_type}"]
    
    mean_WT = adata_WT[:, ].X.mean(0).A1
    mean_KO = adata_KO[:, ].X.mean(0).A1
    
    # Creating a df and selecting the genes is actually faster than selecting them with scanpy
    df = pd.DataFrame({'WT': mean_WT, 'KO':mean_KO}, index=adata_imm.var_names)
    df = df.loc[list_imm_logfold_genes]

    min_val = 0.1

    logfold = np.log2((df['KO'] + min_val) / (df['WT'] + min_val))
    mean = (df['KO'] + df['WT']) / 2
    
    df_logfold.loc[list_imm_logfold_genes, cell_type] = logfold
    df_mean.loc[list_imm_logfold_genes, cell_type] = mean

    df_logfold = df_logfold.astype(float)
    df_mean = df_mean.astype(float)

list_labels = ['Suppressive', 'Stim.', 'Proinf.', 'IL1 proc.', 'Chemoatt.', 'Chemorec.']

pylab.rcParams.update(params)

fig, ax = plt.subplots(1, 1, figsize=(4*1.5, 7*1.5))
text_ax = fig.add_axes([.9, 0.125, .02, 0.755], sharey=ax)
cbar_ax = fig.add_axes([.98, .58, .03, .2])


sns.heatmap(df_logfold, ax=ax, cbar_ax = cbar_ax, cbar=True, cmap='RdBu_r', center=0, 
            yticklabels=True, vmax=4, vmin=-2)

# Plot the categories
counter = 0
x_bar = len(categories_order)
text_ax.tick_params(labelbottom=False, labelleft=False)
text_ax.set_axis_off()

for idx, (category, genes) in enumerate(markers.items()):
    text_ax.plot([0, 0], [counter + 0.25, counter + len(genes) - 0.25], c="#404040")
    text_ax.text(x=0.1, y = counter + len(genes)/2, s=list_labels[idx], rotation=-90, 
                 ha='center', va='center', size='large')
    counter += len(genes)
    
    
plt.savefig(f'results/figures/F3/matrixplot_selected_markers_{date.today().strftime("%Y_%m_%d")}_HEATMAP_LOG2.jpg', dpi=300, bbox_inches="tight")
plt.savefig(f'results/figures/F3/Fig3C.jpg', dpi=300, bbox_inches="tight")

adata_T_cells = adata_imm[adata_imm.obs['subtype_merged_name'].isin(['αβT', 'γδT'])]
adata_T_cells.obs['subtype-condition'] = adata_T_cells.obs['subtype_merged_name'].astype(str) + '-' + adata_T_cells.obs['condition'].astype(str)

t_genes = ['Cd3g', 'Trac', 'Trbc2', 'Cd4', 'Cd8a', 'Foxp3', 'Il2ra', 'Il7r', 'Ctla4', 'Pdcd1', 'Tcf7', 
           'Tox', 'Havcr2', 'Tigit', 'Lag3', 'Gzmb', 'Tnfrsf18', 'Tnfrsf9', 'Tnfrsf4']

adata_sample = adata_T_cells[adata_T_cells.obs['subtype_merged_name'] == 'αβT']
adata_sample = adata_sample[:, t_genes]



X = adata_sample[:, ['Pdcd1', 'Tcf7']].X.toarray()
X_1 = (X[:, 0] * X[:, 1]) + X[:, 0] + X[:, 1]

idx_order = X_1.argsort()[::-1]
idx_order = [i for i in idx_order if i in np.argwhere((X[:, 0] > 0) & (X[:, 1] > 0)).ravel()] + \
            [i for i in idx_order if i in np.argwhere((X[:, 0] > 0) & (X[:, 1] == 0)).ravel()] + \
            [i for i in idx_order if i in np.argwhere((X[:, 0] == 0) & (X[:, 1] > 0)).ravel()] + \
            [i for i in idx_order if i in np.argwhere((X[:, 0] == 0) & (X[:, 1] == 0)).ravel()] 


idx_cells = adata_sample.obs_names[idx_order]

adata_sample = adata_sample[idx_cells]

KO_WT_idx = list(adata_sample[adata_sample.obs['condition'] == 'KO'].obs_names) + list(adata_sample[adata_sample.obs['condition'] == 'WT'].obs_names)

adata_sample = adata_sample[KO_WT_idx]

sc.pl.heatmap(adata_sample, var_names=t_genes, groupby='subtype-condition', use_raw=False, swap_axes=True, show=False)

plt.savefig(f'results/figures/F3/heatmap_T_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")
plt.savefig(f'results/figures/F3/Fig3B.jpg', dpi=300, bbox_inches="tight")

sc.pl.matrixplot(adata_sample, var_names=t_genes, groupby='subtype-condition', use_raw=False, show=False)

plt.savefig(f'results/figures/F3/matrixplot_T_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")

### Analysis os KO vs WT in ab and gd T cells

sc.tl.rank_genes_groups(adata_T_cells, groupby='subtype-condition', groups=['γδT-KO'], 
                       reference='γδT-WT')

for gene in adata_T_cells.uns['rank_genes_groups']['names']['γδT-KO'][:80]:
    sc.pl.umap(adata_T_cells[adata_T_cells.obs['condition'] == 'WT'], 
               color=['subtype-condition', gene], cmap=magma, use_raw=False, ncols=4)
    sc.pl.umap(adata_T_cells[adata_T_cells.obs['condition'] == 'KO'], 
               color=['subtype-condition', gene], cmap=magma, use_raw=False, ncols=4)

genesgd = ['Aldoa', 'Tpi1', 'Eno1', 'Ldha', 'Mif', 'Pgk1', 'Gpi1', 'Itm2c', 'Pgam1', 'Col7a1', 
           'Ypel3', 'Pde7a', 'Usp48', 'Bin1', 'Ssbp2', 'Cflar', 'Slfn2']

sc.pl.matrixplot(adata_T_cells, var_names=genesgd, groupby='subtype-condition', use_raw=False, show=False)
plt.savefig(f'results/figures/F3/matrixplot_DEG_gdT_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")
sc.pl.heatmap(adata_T_cells, var_names=genesgd, groupby='subtype-condition', use_raw=False, swap_axes=True, show=False)
plt.savefig(f'results/figures/F3/heatmap_DEG_gdT_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")


sc.tl.rank_genes_groups(adata_T_cells, groupby='subtype-condition', groups=['αβT-KO'], 
                       reference='αβT-WT')

for gene in adata_T_cells.uns['rank_genes_groups']['names']['αβT-KO'][:80]:
    sc.pl.umap(adata_T_cells[adata_T_cells.obs['condition'] == 'WT'], 
               color=['subtype-condition', gene], cmap=magma, use_raw=False, ncols=4)
    sc.pl.umap(adata_T_cells[adata_T_cells.obs['condition'] == 'KO'], 
               color=['subtype-condition', gene], cmap=magma, use_raw=False, ncols=4)

genesab = ['Aldoa', 'Gpi1', 'Eno1', 'Pgk1', 'Ly6e', 'Tnfrsf9', 'Tpi1', 'Il7r', 
            'Sdf4', 'Shisa5', 'Gmfg', 'Sept1', 'Hif1a', 'Col7a1', 'Tet2', 
            'Stk26', 'Scamp3', 'Trabd', 'Apobec3', 'Vma21', ]

sc.pl.matrixplot(adata_T_cells, var_names=genesab, groupby='subtype-condition', use_raw=False, show=False)
plt.savefig(f'results/figures/F3/matrixplot_DEG_abT_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")
sc.pl.heatmap(adata_T_cells, var_names=genesab, groupby='subtype-condition', use_raw=False, swap_axes=True, show=False)
plt.savefig(f'results/figures/F3/heatmap_DEG_abT_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")


##### Analysis of ab T cells

adata_T_cells.obs['subtype_name'] == 'αβT'

[i for i in adata_T_cells.var_names if 'Havcr' in i]

ab_t_genes = ['Cd3g', 'Trac', 'Trbc2', 'Cd4', 'Cd8a', 'Foxp3', 'Il2ra', 'Il7r', 
              'Pdcd1', 'Tcf7', 'Slamf6', 'Tox', 'Nfatc2', 'Nr4a1', 'Ctla4', 
              'Havcr2', 'Havcr1', 'Tigit', 'Lag3', 'Gzmb', 'Tnfrsf18', 'Tnfrsf9', 'Tnfrsf4', ]

#Havcr4 is not -> add Havcr1 and Havcr2
sc.pl.heatmap(adata_T_cells[adata_T_cells.obs['subtype_name'] == 'αβT'], 
              var_names=ab_t_genes, groupby='subtype-condition', use_raw=False, swap_axes=True, show=False)

plt.savefig(f'results/figures/F3/heatmap_abT_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")

sc.pl.matrixplot(adata_T_cells[adata_T_cells.obs['subtype_name'] == 'αβT'], 
                 var_names=ab_t_genes, groupby='subtype-condition', use_raw=False, show=False)

plt.savefig(f'results/figures/F3/matrixplot_abT_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")

### Analysis of KO vs WT in Neutrophils cells

adata_neu = adata_imm[adata_imm.obs['subtype_merged_name'] == 'Neu']

sc.tl.rank_genes_groups(adata_neu, groupby='subtype_name', groups=['Act. Neu'], 
                       reference='Neu')

for gene in adata_neu.uns['rank_genes_groups']['names']['Act. Neu'][:80]:
    sc.pl.umap(adata_neu[adata_neu.obs['condition'] == 'WT'], 
               color=['subtype_name', gene], cmap=magma, use_raw=False, ncols=4)
    sc.pl.umap(adata_neu[adata_neu.obs['condition'] == 'KO'], 
               color=['subtype_name', gene], cmap=magma, use_raw=False, ncols=4)

genes_neu = ['Ccl3', 'Cxcl2', 'Ier3', 'Nfkbia', 'Hilpda', 
             'Cxcl3', 'Junb', 'Hcar2', 'Gadd45b', 'Ccl4', 'Slpi', 'Clec4d', 'Csta3', 
             'Il1rn', 'Ccrl2', 'Ptgs2', 'Dusp1', 'Il1b', 'Pim1', 'Pnrc1', 
             'Ifrd1', 'Smox', 'Clec4e', 'Zeb2', 'Ankrd33b', 
             'Nlrp3', 'Basp1', 'Ninj1', 'Card19', 'Slc15a3', 'Cd274',  
             'Bcl2a1a', 'Bcl2a1b', 'Acod1', 'Pik3ap1', 
             'Btg1', 'Il1a', 'Ppp1r15a', 'Irak2', 
             'F3', 'Il1f9', 'Tnfaip3', 'Tgif1', 'Marcksl1'] 

adata_sample = adata_neu
adata_sample = adata_sample[:, genes_neu]


idx_cells = adata_sample.obs[['subtype_name', 'condition']].sort_values(by=['subtype_name', 'condition']).index

adata_sample = adata_sample[idx_cells]

adata_X = adata_sample.X.toarray()
adata_X[adata_X > 3] = 3

sc.pl.matrixplot(adata_sample, var_names=genes_neu, groupby='subtype_name', 
                 use_raw=False, show=False, vmax=3)
plt.savefig(f'results/figures/F3/matrixplot_DEG_neu_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")

fig = plt.figure(figsize=(6, 8))
N = 30
axht = fig.add_subplot(N, 1, (3, N-2))
axwt = fig.add_subplot(N, 1, N-1, sharex=axht)
axcell = fig.add_subplot(N, 1, N, sharex=axht)

sns.heatmap(pd.factorize(adata_sample.obs['condition'].values)[0].reshape(1, len(adata_sample)), 
            ax=axwt, cbar=False, cmap=[dict_WT_KO_colors['KO1'], dict_WT_KO_colors['WT1']], 
            yticklabels=False, xticklabels=False)
sns.heatmap(pd.factorize(adata_sample.obs['subtype_name'].values)[0].reshape(1, len(adata_sample)), 
            ax=axcell, cbar=False, cmap=[dict_imm_colors['2: Act. Neu'], dict_imm_colors['3: Neu']], 
            yticklabels=False, xticklabels=False)
sns.heatmap(adata_X.transpose(), ax=axht, cmap='viridis', cbar=False, yticklabels=True, 
            xticklabels=False)

# axht.set_xticklabels(adata_sample.obs_names)
axht.set_yticklabels(adata_sample.var_names, size=8)

axcell.legend(handles=[mpatches.Patch(color=dict_WT_KO_colors['KO1'], label='KO'), 
                       mpatches.Patch(color=dict_WT_KO_colors['WT1'], label='WT'), 
                       mpatches.Patch(color=dict_imm_colors['2: Act. Neu'], label='Act. Neu (Clustr 2)'), 
                       mpatches.Patch(color=dict_imm_colors['3: Neu'], label='Neu (Cluster 3)')], 
               frameon=False, bbox_to_anchor=(0.95, -0.1), ncol=4, fontsize='medium')

plt.savefig(f'results/figures/F3/heatmap_DEG_neu_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")

### Analysis os KO vs WT in Langerhans cells

adata_LC = adata_imm[adata_imm.obs['subtype_merged_name'] == 'LC']

sc.tl.rank_genes_groups(adata_LC, groupby='subtype_name', groups=['Act. LC'], 
                       reference='LC')

for gene in adata_LC.uns['rank_genes_groups']['names']['Act. LC'][:80]:
    sc.pl.umap(adata_LC[adata_LC.obs['condition'] == 'WT'], 
               color=['subtype_name', gene], cmap=magma, use_raw=False, ncols=4)
    sc.pl.umap(adata_LC[adata_LC.obs['condition'] == 'KO'], 
               color=['subtype_name', gene], cmap=magma, use_raw=False, ncols=4)

genes_LC = ['Actg1', 'Ccr7', 'Bcl2a1b', 'Myl6', 'Crip1', 
            'Ccl22', 'Samsn1',   'Gng5', 'Il4i1', 'Tspo', 'Psmb8', 'Arpc2', 'Abracl', 'Fscn1',     
            'Bcl2a1a', 'Bcl2a1d', 'Arhgdia', 'Coro1a', 'Ostf1',    
            'Plxnd1', 'Selplg',  
            'Cacnb3', 'Wdr1', 'Iscu', 'Cd200', 'Jak2', 'Psme2', 'Ramp3', 
            'Tmem123', 'Necap2', 'Sri', 'Wnk1', 'Sin3b', 'Tpm4', 'Myo1g', 'Stk4', 'Serpinb9b', 'Sept9', 
            'St8sia4', 'Cyfip1', 'Psme1', 'Rassf2', 'Ppdpf'
             ]

adata_sample = adata_LC
adata_sample = adata_sample[:, genes_LC]


idx_cells = adata_sample.obs[['subtype_name', 'condition']].sort_values(by=['subtype_name', 'condition']).index

adata_sample = adata_sample[idx_cells]

adata_X = adata_sample.X.toarray()
adata_X[adata_X > 3] = 3

sc.pl.matrixplot(adata_sample, var_names=genes_LC, groupby='subtype_name', 
                 use_raw=False, show=False, vmax=3)
plt.savefig(f'results/figures/F3/matrixplot_DEG_LC_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")

fig = plt.figure(figsize=(6, 8))
N = 30
axht = fig.add_subplot(N, 1, (3, N-2))
axwt = fig.add_subplot(N, 1, N-1, sharex=axht)
axcell = fig.add_subplot(N, 1, N, sharex=axht)

sns.heatmap(pd.factorize(adata_sample.obs['condition'].values)[0].reshape(1, len(adata_sample)), 
            ax=axwt, cbar=False, cmap=[dict_WT_KO_colors['KO1'], dict_WT_KO_colors['WT1']], 
            yticklabels=False, xticklabels=False)
sns.heatmap(pd.factorize(adata_sample.obs['subtype_name'].values)[0].reshape(1, len(adata_sample)), 
            ax=axcell, cbar=False, cmap=[dict_imm_colors['4: Act. LC'], dict_imm_colors['5: LC']], 
            yticklabels=False, xticklabels=False)
sns.heatmap(adata_X.transpose(), ax=axht, cmap='viridis', cbar=False, yticklabels=True, 
            xticklabels=False)

# axht.set_xticklabels(adata_sample.obs_names)
axht.set_yticklabels(adata_sample.var_names, size=8)

axcell.legend(handles=[mpatches.Patch(color=dict_WT_KO_colors['KO1'], label='KO'), 
                       mpatches.Patch(color=dict_WT_KO_colors['WT1'], label='WT'), 
                       mpatches.Patch(color=dict_imm_colors['4: Act. LC'], label='Act. LC (Cluster 4)'), 
                       mpatches.Patch(color=dict_imm_colors['5: LC'], label='LC (Cluster 5)')], 
               frameon=False, bbox_to_anchor=(1.05, -0.1), ncol=4, fontsize='medium')

plt.savefig(f'results/figures/F3/heatmap_DEG_LC_markers_{date.today().strftime("%Y_%m_%d")}.jpg', dpi=300, bbox_inches="tight")

### Save adata

adata_imm.write_h5ad('adatas/adata_immune.h5', compression='gzip')

os.system('jupyter nbconvert --to html 4_figures_immune.ipynb')
