==> 6_ligand_receptor_interaction.ipynb <==
# Figures associated to LR interactions

**You need to run notebooks 1 to 5, and D2 first!!!!**

from cellassign import assign_cats
import gzip
import itertools as itl
import matplotlib.pyplot as plt
import matplotlib as mpl
import numpy as np
import os
import pandas as pd
import scanpy as sc
import scanpy.external as sce
import seaborn as sns
import triku as tk
import itertools
import matplotlib.pylab as pylab

from functions import adata_plot_KOvsWT, stat_annot_gene, save_adata
from functions import dict_WT_KO_colors

# Palettes for UMAP gene expression

magma = [plt.get_cmap('magma')(i) for i in np.linspace(0,1, 80)]
magma[0] = (0.88, 0.88, 0.88, 1)
magma = mpl.colors.LinearSegmentedColormap.from_list("", magma[:65])

params = {'figure.dpi': 200,
          'legend.fontsize': 'large',
         'axes.labelsize': 'x-large',
         'axes.titlesize':'x-large',
         'xtick.labelsize':'x-large',
         'ytick.labelsize':'x-large', 
         'axes.grid': False}
pylab.rcParams.update(params)
# mpl.style.use('default')

pd.set_option('display.max_columns', None)

## Functions and variables

%store -r dict_WT_KO_colors
%store -r seed

## Load adata

adata_fb = sc.read('adatas/adata_fibroblast.h5')
adata_krt = sc.read('adatas/adata_keratinocyte.h5')
adata_imm = sc.read('adatas/adata_immune.h5')
adata_vasc = sc.read('adatas/adata_vasculature.h5')
adata_all_harmony = sc.read_h5ad('adatas/adata_all_harmony.h5')

adata_combined = sc.AnnData.concatenate(adata_fb, adata_krt, adata_imm, adata_vasc)
adata_combined.obs_names = [i[:-2] for i in adata_combined.obs_names]
# We do this because some gene expression may be lost when removing genes in subtype anndatas
# adata_combined.X = adata_all_harmony[adata_combined.obs_names, adata_combined.var_names].X  

print([(i[1], i[0]) for i in [('Ccl4', 'Ccr5'), ('Ccl6', 'Ccr1'), ('Ccl7', 'Ccr1'), ('Ccl7', 'Ccr2'),
            ('Ccl8', 'Ccr3'), ('Ccl22', 'Ccr4'),]])


LR_pairs = [('C3ar1', 'C3'), ('Csf1', 'Csf1r'), ('Csf2', 'Csf2ra'), ('Cxcl1', 'Cxcr2'), ('Cxcl2', 'Cxcr4'), 
            ('Cxcr6', 'Cxcl16'), ('Ccr2', 'Ccl2'), ('Ccr4', 'Ccl2'), ('Ccr1', 'Ccl3'), ('Ccr4', 'Ccl3'),  ('Ccr5', 'Ccl3'), 
            ('Ccr5', 'Ccl4'), ('Ccr1', 'Ccl6'), ('Ccr1', 'Ccl7'), ('Ccr2', 'Ccl7'), # ('Ccr3', 'Ccl8'), ('Ccr4', 'Ccl22'), 
            ('Il1f9', 'Il1rl2'), ('Il1a', 'Il1rl1'), ('Il1b', 'Il1rl1'),
            ('Il33', 'Il1rl1'), ('Tnf', 'Tnfrsf1b'), ('Tnfsf9', 'Tnfrsf9'), ('Vegfa', 'Flt1'), ('Vegfb', 'Flt1'),
            ('Pdcd1', 'Cd274'),]

# for pair in LR_pairs:
#     print(pair[0], pair[0] in adata_imm.var_names, pair[0] in adata_rest.var_names,)
#     print(pair[1], pair[1] in adata_imm.var_names, pair[1] in adata_rest.var_names, '\n\n')

def plot_violin_ab(gene_1, gene_2, adata_1, adata_2):
    fig = plt.figure(constrained_layout=False, figsize=(12, 4))
    gs1 = fig.add_gridspec(nrows=1, ncols=4)
    ax1, ax2 = fig.add_subplot(gs1[0]), fig.add_subplot(gs1[1:])

    sc.pl.violin(adata_1, groupby='subtype', keys=gene_1, rotation=90, ax=ax1, show=False, use_raw=False)
    sc.pl.violin(adata_2, groupby='subtype', keys=gene_2, rotation=90, 
                 ax=ax2, show=False, use_raw=False)
    plt.tight_layout()

adata_rest = sc.AnnData.concatenate(adata_fb, adata_krt, adata_vasc)

for pair_1, pair_2 in LR_pairs:
    plot_violin_ab(pair_1, pair_2, adata_imm, adata_rest)

os.system('jupyter nbconvert --to html 6_ligand_receptor_interaction.ipynb')
